* Fixed issues with node packages being pulled into geometry extraction, which meant they were coming into share.
