+++ src/ifc/ifc_model_geometry.ts
@@ -2,17 +2,12 @@
 import { CanonicalMesh, CanonicalMeshType } from '../core/canonical_mesh'
 import { ModelGeometry } from '../core/model'
 import IfcStepModel from './ifc_step_model'
-import path from 'path'
-import fsPromises from 'fs/promises'
 import { IfcBooleanResult, IfcGeometricRepresentationItem } from './ifc4_gen'
-import crypto from 'crypto'
-import { CanonicalMaterial, dumpMTL, getMTLCleanName } from '../core/canonical_material'
+import { CanonicalMaterial, getMTLCleanName } from '../core/canonical_material'
 
-
-const MAX_FILES_OPEN = 64
 
 /**
- *
+ * Geometry for an IFC model.
  */
 export class IfcModelGeometry implements ModelGeometry {
 
@@ -169,73 +164,5 @@ export class IfcModelGeometry implements ModelGeometry {
       yield [geometryItem, objFileContents, undefined, undefined]
     }
   }
-
-  /**
-   * Dump the OBJs in this to a particular folder
-   *
-   * @param folder The folder to dump to
-   * @return {Promise<void>} A promise to wait on when this completes.
-   */
-  public async dumpOBJs( folder: string ): Promise< void > {
-
-    await fsPromises.mkdir( folder, { recursive: true })
-
-    const writePromises: Promise< void >[] = []
-
-    for ( const [geometryItem, objFileContents, material, materialName] of this.objs() ) {
-
-      const localID = geometryItem.localID
-
-      const outputExpressID = geometryItem.expressID
-      const outputFileName =
-        outputExpressID !== void 0 ?
-          `${outputExpressID}.obj` :
-          `${localID}_inline.obj`
-
-      const RADIX_CHARS = 2
-
-      const outputFolder = path.join(
-          folder, outputExpressID !== void 0 ?
-            String( outputExpressID ).padStart( RADIX_CHARS, '0' ).substring( 0, RADIX_CHARS ) :
-            'inline' )
-
-      await fsPromises.mkdir( outputFolder, { recursive: true } )
-
-      if ( material !== void 0 ) {
-
-        const materialFilePath = path.join( outputFolder, materialName )
-
-        writePromises.push( fsPromises.writeFile( materialFilePath, dumpMTL( material ) ) )
-      }
-
-      const outputFilePath = path.join( outputFolder, outputFileName )
-
-      writePromises.push( fsPromises.writeFile( outputFilePath, objFileContents ) )
-
-      if ( writePromises.length >= MAX_FILES_OPEN ) {
-
-        await Promise.all( writePromises )
-        writePromises.length = 0
-      }
-    }
-
-    await Promise.all( writePromises )
-  }
-
-  /**
-   * Build a set of hashes with their matching IFC curves.
-   *
-   * @yields {[IfcGeometricRepresentationItem, Uint8Array]} A list of curves with
-   * their corresponding hash.
-   */
-  public* hashes(): IterableIterator< [IfcGeometricRepresentationItem, Uint8Array] >  {
-
-    for ( const [curveItem, objFileContents] of this.objs() ) {
-
-      const objHash = crypto.createHash( 'sha1' ).update( objFileContents ).digest()
-
-      yield [curveItem, objHash]
-    }
-  }
-
 }
+
