+++ src/ifc/ifc_model_profile.ts
@@ -3,13 +3,8 @@
 import { CanonicalProfile } from '../core/canonical_profile'
 import {ModelProfile} from '../core/model'
 import { IfcProfileDef } from './ifc4_gen'
-import path from 'path'
-import fsPromises from 'fs/promises'
-import crypto from 'crypto'
 import IfcStepModel from './ifc_step_model'
 
-
-const MAX_FILES_OPEN = 64
 
 /**
  * IFC model profile cache.
@@ -87,65 +82,4 @@ export class IfcModelProfile implements ModelProfile {
     }
 
   }
-
-  /**
-   * Dump the OBJs in this to a particular folder
-   *
-   * @param folder The folder to dump to
-   * @return {Promise<void>} A promise to wait on when this completes.
-   */
-  public async dumpOBJs( folder: string ): Promise< void > {
-
-    await fsPromises.mkdir( folder, { recursive: true })
-
-    const writePromises: Promise< void >[] = []
-
-    for ( const [profileItem, objFileContents] of this.objs() ) {
-
-      const localID = profileItem.localID
-
-      const outputExpressID = profileItem.expressID
-      const outputFileName =
-      outputExpressID !== void 0 ?
-          `${outputExpressID}.obj` :
-          `${localID}_inline.obj`
-
-      const RADIX_CHARS = 2
-
-      const outputFolder = path.join(
-          folder, outputExpressID !== void 0 ?
-            String( outputExpressID ).padStart( RADIX_CHARS, '0' ).substring( 0, RADIX_CHARS ) :
-            'inline' )
-
-      await fsPromises.mkdir( outputFolder, { recursive: true } )
-
-      const outputFilePath = path.join( outputFolder, outputFileName )
-
-      writePromises.push( fsPromises.writeFile( outputFilePath, objFileContents ) )
-
-      if ( writePromises.length >= MAX_FILES_OPEN ) {
-
-        await Promise.all( writePromises )
-        writePromises.length = 0
-      }
-    }
-
-    await Promise.all( writePromises )
-  }
-
-  /**
-   * Build a set of hashes with their matching IFC curves.
-   *
-   * @yields {[IfcProfileDef, Uint8Array]} A list of curves with their corresponding hash.
-   */
-  public* hashes(): IterableIterator< [IfcProfileDef, Uint8Array] >  {
-
-    for ( const [profileItem, objFileContents] of this.objs() ) {
-
-      const objHash = crypto.createHash( 'sha1' ).update( objFileContents ).digest()
-
-      yield [profileItem, objHash]
-    }
-  }
-
 }
